requirements.txt
asgiref==3.9.0
Django==5.2.4
sqlparse==0.5.3

opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0

djangorestframework==3.15.2
torch==2.2.2
torchvision==0.17.2
facenet-pytorch==2.6.0
scikit-learn==1.5.1
seaborn==0.13.2
matplotlib==3.9.0

